Move `#[cfg(test)]` methods into `tracker::tests` module
